requirements.txt
gradio
pandas
numpy
matplotlib
requests
anthropic
modal
yfinance
python-dotenv
openai